sms-reminder-system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
twilio==8.10.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary
